Fix readthedocs build
Seems like there's some funny business (as always) in the world of
python packaging. There is a very helpful thread on readthedocs github:

https://github.com/readthedocs/readthedocs.org/issues/4912

Looking through some other people who have fought through this, the main
issues seem to be that:

  * There's no real "pip" notion of a "dev-dependency",
  * Poetry doesn't yet support dependency groups other than
    "dependencies" and "dev-dependencies", and
  * You need to utilize a special incantation in `.readthedocs.yml` to
    `pip install -e .` the current package.

This commit responds to all these issues. Thanks especially to this
super helpful commit basically showing the way:

https://github.com/TeoZosa/structlog-sentry-logger/commit/f4913dcb3c39bb1b4cf7247a0ac7641aa31f35f9

I've tested this on a fake repo setup just to test rtd:

https://testxxxxxtest.readthedocs.io/en/latest/

This seesm to be working. :-)